Add NotFoundPage component and error handling
Created a NotFoundPage component to display whenever a page is not found or an error occurs. Integrated error handling in Articles, Article, and utils functions to catch errors, update relevant state and display NotFoundPage with the associated error message. Additionally, NotFoundPage is set to be the default for any non-defined routes in the app.